# Reuse string buffer in stringTuples.Swap