This is a broken down piece of #627

[Don't merge until Prometheus 2.11 is out]